opnsense-bridge-api/requirements.txt
# Core dependencies for FastAPI application
fastapi==0.95.0
uvicorn[standard]==0.20.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiohttp==3.8.4
pyyaml==6.0
gunicorn==20.1.0

# Notification dependencies
aiosmtplib==2.0.0
requests==2.28.2

# Monitoring dependencies
prometheus-client==0.16.0

# Testing dependencies
pytest==7.2.2
pytest-asyncio==0.20.3
httpx==0.23.3
pytest-cov==4.0.0
mock==5.0.1